Adds support for Nova assisted volume snapshots
This change will allow creating/deleting in-use volumes exposed
by the SMB Cinder driver. This includes snapshotting volume backed
instances.

We're suspending the instance and creating a temporary instance
snapshot in case of IDE disks (which cannot be detached), while
just pausing it when snapshotting SCSI disks.

In the future, we're going to support VHDSet images. Nova will
not have to assist snapshotting such images. Still, we need this
change as VHDSet images are only supported from Windows Server 2016
onwards and have to reside on CSVs.

Depends-On: I58e876ca63d0f8baf46de58ed18fdc98285f30d5

Co-Authored-By: Lucian Petrut <lpetrut@cloudbasesolutions.com>

Change-Id: I6b4a1e4c806c5f81621119acc713c1092d3afb78
Implements: blueprint hyperv-assisted-volume-snapshot