Use LibvirtLiveMigrateData
During live migration, a migrate data object is passed around the
compute nodes and the scheduler. This is used in order to determine
if shared storage is used, the manager deciding whether or not
cleanup is required while the driver deciding whether it should
move files.

The Mitaka version of the HyperVLiveMigrateData object does not
contain the field we need, so we're going to do a little hack and
use the LibvirtLiveMigrateData object instead.

We're passing the 'is_shared_block_storage' field as well, as the
manager will act based on it, depending on the migrate object type.