Update device metadata when attaching disks/NICs
This change ensures that we're updating the instance device metadata
when attaching or detaching instance volumes or NICs.

Unlike the other Nova drivers, we're including all Cinder volumes
in the device metadata, not just the tagged ones. The reason is that
this allows Hyper-V guests to properly identify attached disk paths.

Note that the disk mountpoints (e.g. /dev/sdc) are ignored by the
Hyper-V driver. Also, we cannot override SCSI identifiers, unlike
the Libvirt driver, which will use volume IDs as SCSI IDs.

Closes-Bug: #1723431

Depends-On: I7a4c242c80b5b3562bf9513deb89071b5bb5bf28
Depends-On: I6f2c6ecf71d3254d64bc24574619c18184d3d055

Change-Id: I6853b1d167fe675435f3fb8f1fdb0cf0b63bed65